+++ apps/studio/data/replication/create-destination-pipeline-mutation.ts
@@ -2,8 +2,9 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import type { components } from 'api-types'
 import { toast } from 'sonner'
 
-import { optionalSecret } from './destination-secret-utils'
 import { replicationKeys } from './keys'
+import type { DucklakeDestinationConfig } from './types'
+import { isDucklakeSupabaseConfig } from './utils'
 import { handleError, post } from '@/data/fetchers'
 import type { ResponseError, UseCustomMutationOptions } from '@/types'
 
@@ -11,14 +12,9 @@ type CreateDestinationPipelineBody =
   components['schemas']['CreateReplicationDestinationPipelineBody']
 type CreateDestinationApiConfig = CreateDestinationPipelineBody['destination_config']
 type CreatePipelineApiConfig = CreateDestinationPipelineBody['pipeline_config']
-type UpdateDestinationPipelineBody =
-  components['schemas']['UpdateReplicationDestinationPipelineBody']
-type UpdateDestinationApiConfig = UpdateDestinationPipelineBody['destination_config']
 
 type CreateBigQueryApiConfig = Extract<CreateDestinationApiConfig, { big_query: unknown }>
-type UpdateBigQueryApiConfig = Extract<UpdateDestinationApiConfig, { big_query: unknown }>
 type CreateDucklakeApiConfig = Extract<CreateDestinationApiConfig, { ducklake: unknown }>
-type UpdateDucklakeApiConfig = Extract<UpdateDestinationApiConfig, { ducklake: unknown }>
 
 export type DestinationConfig =
   | { bigQuery: BigQueryDestinationConfig }
@@ -45,42 +41,6 @@ export type IcebergDestinationConfig = {
   s3Region: string
 }
 
-// "Custom parameters" DuckLake: caller provides the PostgreSQL catalog URL and the
-// S3-compatible storage credentials directly.
-export type DucklakeManualDestinationConfig = {
-  catalogUrl: string
-  dataPath: string
-  poolSize?: number
-  s3AccessKeyId: string
-  s3SecretAccessKey: string
-  s3Region: string
-  s3Endpoint: string
-  s3UrlStyle?: 'path' | 'vhost'
-  s3UseSsl?: boolean
-  metadataSchema?: string
-}
-
-// "Use Supabase" DuckLake: caller provides Supabase project refs and a bucket; the platform
-// API resolves these into a catalog URL + provisioned S3 credentials before persisting.
-export type DucklakeSupabaseDestinationConfig = {
-  catalogProjectRef: string
-  storageProjectRef: string
-  bucket: string
-  path?: string
-  poolSize?: number
-  metadataSchema?: string
-}
-
-export type DucklakeDestinationConfig =
-  | DucklakeManualDestinationConfig
-  | DucklakeSupabaseDestinationConfig
-
-function isDucklakeSupabaseConfig(
-  config: DucklakeDestinationConfig
-): config is DucklakeSupabaseDestinationConfig {
-  return 'catalogProjectRef' in config
-}
-
 // Maps the studio-side BigQuery config to the snake_case `{ big_query: ... }` payload accepted
 // by the platform API. Shared by the create and validate mutations.
 export function buildBigQueryApiConfig(config: BigQueryDestinationConfig): CreateBigQueryApiConfig {
@@ -89,20 +49,6 @@ export function buildBigQueryApiConfig(config: BigQueryDestinationConfig): Creat
       project_id: config.projectId,
       dataset_id: config.datasetId,
       service_account_key: config.serviceAccountKey,
-      connection_pool_size: config.connectionPoolSize,
-      max_staleness_mins: config.maxStalenessMins,
-    },
-  }
-}
-
-export function buildBigQueryUpdateApiConfig(
-  config: BigQueryDestinationConfig
-): UpdateBigQueryApiConfig {
-  return {
-    big_query: {
-      project_id: config.projectId,
-      dataset_id: config.datasetId,
-      service_account_key: optionalSecret(config.serviceAccountKey),
       connection_pool_size: config.connectionPoolSize,
       max_staleness_mins: config.maxStalenessMins,
     },
@@ -140,44 +86,6 @@ export function buildDucklakeApiConfig(config: DucklakeDestinationConfig): Creat
       pool_size: config.poolSize,
       s3_access_key_id: config.s3AccessKeyId,
       s3_secret_access_key: config.s3SecretAccessKey,
-      s3_region: config.s3Region,
-      s3_endpoint: config.s3Endpoint,
-      s3_url_style: config.s3UrlStyle,
-      s3_use_ssl: config.s3UseSsl,
-      metadata_schema: config.metadataSchema,
-    },
-  }
-}
-
-export function buildDucklakeUpdateApiConfig(
-  config: DucklakeDestinationConfig
-): UpdateDucklakeApiConfig {
-  if (isDucklakeSupabaseConfig(config)) {
-    return {
-      ducklake: {
-        catalog: {
-          type: 'supabase_project',
-          project_ref: config.catalogProjectRef,
-          pool_size: config.poolSize,
-          metadata_schema: config.metadataSchema,
-        },
-        storage: {
-          type: 'supabase_storage',
-          project_ref: config.storageProjectRef,
-          bucket: config.bucket,
-          ...(config.path ? { path: config.path } : {}),
-        },
-      },
-    }
-  }
-
-  return {
-    ducklake: {
-      catalog_url: optionalSecret(config.catalogUrl),
-      data_path: config.dataPath,
-      pool_size: config.poolSize,
-      s3_access_key_id: optionalSecret(config.s3AccessKeyId),
-      s3_secret_access_key: optionalSecret(config.s3SecretAccessKey),
       s3_region: config.s3Region,
       s3_endpoint: config.s3Endpoint,
       s3_url_style: config.s3UrlStyle,
@@ -302,79 +210,6 @@ export const buildCreateDestinationApiConfig = (
     const { url, user, password, database, engine } = destinationConfig.clickHouse
 
     return { clickhouse: { url, user, password, database, engine } }
-  }
-
-  throw new Error(
-    'Invalid destination config: must specify bigQuery, iceberg, ducklake, snowflake, or clickHouse'
-  )
-}
-
-export const buildUpdateDestinationApiConfig = (
-  destinationConfig: DestinationConfig
-): UpdateDestinationApiConfig => {
-  if ('bigQuery' in destinationConfig) {
-    return buildBigQueryUpdateApiConfig(destinationConfig.bigQuery)
-  }
-
-  if ('iceberg' in destinationConfig) {
-    const {
-      projectRef,
-      warehouseName,
-      namespace,
-      catalogToken,
-      s3AccessKeyId,
-      s3SecretAccessKey,
-      s3Region,
-    } = destinationConfig.iceberg
-
-    return {
-      iceberg: {
-        supabase: {
-          project_ref: projectRef,
-          warehouse_name: warehouseName,
-          namespace,
-          catalog_token: optionalSecret(catalogToken),
-          s3_access_key_id: optionalSecret(s3AccessKeyId),
-          s3_secret_access_key: optionalSecret(s3SecretAccessKey),
-          s3_region: s3Region,
-        },
-      },
-    }
-  }
-
-  if ('ducklake' in destinationConfig) {
-    return buildDucklakeUpdateApiConfig(destinationConfig.ducklake)
-  }
-
-  if ('snowflake' in destinationConfig) {
-    const { accountId, user, privateKey, privateKeyPassphrase, database, schema, role } =
-      destinationConfig.snowflake
-
-    return {
-      snowflake: {
-        account_id: accountId,
-        user,
-        private_key: optionalSecret(privateKey),
-        private_key_passphrase: optionalSecret(privateKeyPassphrase),
-        database,
-        schema,
-        role,
-      },
-    }
-  }
-
-  if ('clickHouse' in destinationConfig) {
-    const { url, user, password, database, engine } = destinationConfig.clickHouse
-
-    return {
-      clickhouse: {
-        url,
-        user,
-        password: optionalSecret(password),
-        database,
-        engine,
-      },
-    }
   }
 
   throw new Error(
