+++ apps/studio/data/replication/types.ts
@@ -1,0 +1,29 @@
+// "Custom parameters" DuckLake: caller provides the PostgreSQL catalog URL and the
+// S3-compatible storage credentials directly.
+export type DucklakeManualDestinationConfig = {
+  catalogUrl: string
+  dataPath: string
+  poolSize?: number
+  s3AccessKeyId: string
+  s3SecretAccessKey: string
+  s3Region: string
+  s3Endpoint: string
+  s3UrlStyle?: 'path' | 'vhost'
+  s3UseSsl?: boolean
+  metadataSchema?: string
+}
+
+// "Use Supabase" DuckLake: caller provides Supabase project refs and a bucket; the platform
+// API resolves these into a catalog URL + provisioned S3 credentials before persisting.
+export type DucklakeSupabaseDestinationConfig = {
+  catalogProjectRef: string
+  storageProjectRef: string
+  bucket: string
+  path?: string
+  poolSize?: number
+  metadataSchema?: string
+}
+
+export type DucklakeDestinationConfig =
+  | DucklakeManualDestinationConfig
+  | DucklakeSupabaseDestinationConfig
