+++ apps/studio/data/replication/utils.ts
@@ -1,3 +1,4 @@
+import type { DucklakeDestinationConfig, DucklakeSupabaseDestinationConfig } from './types'
 import { MAX_RETRY_FAILURE_COUNT } from '@/data/query-client'
 import { ResponseError } from '@/types'
 
@@ -35,4 +36,10 @@ export const checkReplicationFeatureFlagRetry = (
   }
 
   return false
+}
+
+export function isDucklakeSupabaseConfig(
+  config: DucklakeDestinationConfig
+): config is DucklakeSupabaseDestinationConfig {
+  return 'catalogProjectRef' in config
 }
